fix: change dir
