+++ packages/core/client/src/data-source/collection-fields-to-initializer-items/CollectionFieldsToFormInitializerItems.tsx
