+++ packages/core/client/src/data-source/collection-fields-to-initializer-items/items/AssociationCollectionFields.tsx
@@ -1,20 +1,20 @@
-import React, { FC } from "react";
+import React, { FC } from 'react';
 
-import { InheritanceCollectionMixin } from "../../../collection-manager";
-import { AssociationCollectionFieldsProps, getInitializerItemsByFields } from "../utils";
-import { SchemaInitializerChildren, SchemaInitializerItemGroup, SchemaInitializerItemType } from "../../../application/schema-initializer";
+import { InheritanceCollectionMixin } from '../../../collection-manager';
+import { AssociationCollectionFieldsProps, getInitializerItemsByFields } from '../utils';
+import {
+  SchemaInitializerChildren,
+  SchemaInitializerItemGroup,
+  SchemaInitializerItemType,
+} from '../../../application/schema-initializer';
 
 export const AssociationCollectionFields: FC<AssociationCollectionFieldsProps> = (props) => {
-  const {
-    filterAssociationField,
-    filterSelfField = () => true,
-    getSchema,
-    ...otherProps } = props;
+  const { filterAssociationField, filterSelfField = () => true, getSchema, ...otherProps } = props;
   const { collection, t, collectionManager } = props.context;
   const fields = collection.getFields();
   const associationInterfaces = ['o2o', 'oho', 'obo', 'm2o']; // 关联字段类型
   const associationFields = fields
-    .filter(field => {
+    .filter((field) => {
       return associationInterfaces.includes(field.interface);
     })
     .filter((field) => filterSelfField(field, props.context));
@@ -24,7 +24,9 @@ export const AssociationCollectionFields: FC<AssociationCollectionFieldsProps> =
   const children = associationFields
     .map((associationField) => {
       // 获取关联表
-      const associationCollection = collectionManager.getCollection<InheritanceCollectionMixin>(associationField.target!)!;
+      const associationCollection = collectionManager.getCollection<InheritanceCollectionMixin>(
+        associationField.target!,
+      )!;
       if (!associationCollection) return null;
       // 获取父表
       const associationCollectionFields = associationCollection?.getAllFields();
@@ -37,7 +39,7 @@ export const AssociationCollectionFields: FC<AssociationCollectionFieldsProps> =
       const newContext = {
         ...props.context,
         collection: associationCollection,
-      }
+      };
 
       const getAssociationFieldSchema: AssociationCollectionFieldsProps['getSchema'] = (field, context) => {
         const schema = getSchema(field, context);
@@ -46,22 +48,26 @@ export const AssociationCollectionFields: FC<AssociationCollectionFieldsProps> =
           'x-read-pretty': true,
           name: `${associationField.name}.${field.name}`,
           'x-collection-field': `${collection.name}.${associationField.name}.${field.name}`,
-        }
-      }
+        };
+      };
 
       return {
         type: 'subMenu',
         name: associationField.uiSchema?.title,
         title: associationField.uiSchema?.title,
-        children: getInitializerItemsByFields({
-          ...otherProps,
-          filter: filterAssociationField,
-          getSchema: getAssociationFieldSchema,
-        }, associationCollectionFields!, newContext),
+        children: getInitializerItemsByFields(
+          {
+            ...otherProps,
+            filter: filterAssociationField,
+            getSchema: getAssociationFieldSchema,
+          },
+          associationCollectionFields!,
+          newContext,
+        ),
       } as SchemaInitializerItemType;
-    })
+    });
 
   if (!children.length) return null;
 
-  return <SchemaInitializerItemGroup title={t('new | Display association fields')} children={children} />
-}
+  return <SchemaInitializerItemGroup title={t('Display association fields')}>{children}</SchemaInitializerItemGroup>;
+};
