+++ packages/core/client/src/data-source/collection-fields-to-initializer-items/utils/type.ts
