+++ packages/core/client/src/data-source/index.ts
@@ -16,4 +16,4 @@ export * from './data-block';
 export * from './data-source';
 export * from './collection-record';
 export * from './utils';
-export * from './collection-fields-initializer-items';
+export * from './collection-fields-to-initializer-items';
